requirements.txt
sqlalchemy
pdfplumber
python-docx
pandas
python-dotenv
google-cloud-orchestration-airflow
google-cloud-bigquery
google-cloud-storage